+++ src/main/java/com/eternalcoders/pointedge/service/EmployeeService.java
@@ -2,14 +2,62 @@ package com.eternalcoders.pointedge.service;
 
 import com.eternalcoders.pointedge.dto.EmployeeDTO;
 import com.eternalcoders.pointedge.entity.Employee;
+import com.eternalcoders.pointedge.exception.ResourceNotFoundException;
+import com.eternalcoders.pointedge.repository.AttendanceRepository;
+import com.eternalcoders.pointedge.repository.EmployeeRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
 import java.util.List;
 
-public interface EmployeeService {
-    List<Employee> getAllEmployees();
-    Employee getEmployeeById(Long id);
-    Employee createEmployee(EmployeeDTO employeeDTO);
-    Employee updateEmployee(Long id, EmployeeDTO employeeDTO);
-    void deleteEmployee(Long id);
-    List<Employee> searchEmployees(String query);
-    List<Employee> findByStatus(Employee.EmployeeStatus status);
+@Service
+@Transactional
+public class EmployeeService {
+
+    @Autowired
+    private EmployeeRepository employeeRepository;
+
+    @Autowired
+    private AttendanceRepository attendanceRepository;
+
+    public List<Employee> getAllEmployees() {
+        return employeeRepository.findAll();
+    }
+
+    public Employee getEmployeeById(Long id) {
+        return employeeRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Employee not found with id: " + id));
+    }
+
+    public Employee createEmployee(EmployeeDTO employeeDTO) {
+        Employee employee = new Employee();
+        employee.setName(employeeDTO.getName());
+        employee.setRole(employeeDTO.getRole());
+        employee.setStatus(employeeDTO.getStatus());
+        employee.setAvatar(employeeDTO.getAvatar());
+        return employeeRepository.save(employee);
+    }
+
+    public Employee updateEmployee(Long id, EmployeeDTO employeeDTO) {
+        Employee employee = getEmployeeById(id);
+        employee.setName(employeeDTO.getName());
+        employee.setRole(employeeDTO.getRole());
+        employee.setStatus(employeeDTO.getStatus());
+        employee.setAvatar(employeeDTO.getAvatar());
+        return employeeRepository.save(employee);
+    }
+
+    public void deleteEmployee(Long id) {
+        Employee employee = getEmployeeById(id);
+        employeeRepository.delete(employee);
+    }
+
+    public List<Employee> searchEmployees(String query) {
+        return employeeRepository.findByNameContainingIgnoreCase(query);
+    }
+
+    public List<Employee> findByStatus(Employee.EmployeeStatus status) {
+        return employeeRepository.findByStatus(status);
+    }
 }
