--- src/main/java/com/eternalcoders/pointedge/service/EmployeeServiceImpl.java
@@ -1,71 +1,0 @@
-package com.eternalcoders.pointedge.service;
-
-import com.eternalcoders.pointedge.dto.EmployeeDTO;
-import com.eternalcoders.pointedge.entity.Employee;
-import com.eternalcoders.pointedge.exception.ResourceNotFoundException;
-import com.eternalcoders.pointedge.repository.AttendanceRepository;
-import com.eternalcoders.pointedge.repository.EmployeeRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-public class EmployeeServiceImpl implements EmployeeService {
-
-    @Autowired
-    private EmployeeRepository employeeRepository;
-
-     @Autowired
-    private AttendanceRepository attendanceRepository;
-
-    
-
-    @Override
-    public List<Employee> getAllEmployees() {
-        return employeeRepository.findAll();
-    }
-
-    @Override
-    public Employee getEmployeeById(Long id) {
-        return employeeRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Employee not found with id: " + id));
-    }
-
-    @Override
-    public Employee createEmployee(EmployeeDTO employeeDTO) {
-        Employee employee = new Employee();
-        employee.setName(employeeDTO.getName());
-        employee.setRole(employeeDTO.getRole());
-        employee.setStatus(employeeDTO.getStatus()); // Already enum
-        employee.setAvatar(employeeDTO.getAvatar());
-        return employeeRepository.save(employee);
-    }
-    
-    @Override
-    public Employee updateEmployee(Long id, EmployeeDTO employeeDTO) {
-        Employee employee = getEmployeeById(id);
-        employee.setName(employeeDTO.getName());
-        employee.setRole(employeeDTO.getRole());
-        employee.setStatus(employeeDTO.getStatus()); // Already enum
-        employee.setAvatar(employeeDTO.getAvatar());
-        return employeeRepository.save(employee);
-    }
-    
-    @Override
-    public void deleteEmployee(Long id) {
-        Employee employee = getEmployeeById(id);
-        employeeRepository.delete(employee);
-    }
-    
-    @Override
-    public List<Employee> searchEmployees(String query) {
-        return employeeRepository.findByNameContainingIgnoreCase(query);
-    }
-    
-    @Override
-    public List<Employee> findByStatus(Employee.EmployeeStatus status) {
-        return employeeRepository.findByStatus(status);
-    }
-    
-}
